requirements.txt
Flask==3.0.3
langchain==0.1.20
langchain-community==0.0.38
langchain_core==0.1.52
langchain_openai==0.0.2.post1
langchain_pinecone==0.1.3
langchain_text_splitters==0.0.2
numpy==1.23.5
openai==1.14.3
pinecone_client==5.0.1
pydantic==2.8.2
pymongo==4.10.1
Requests==2.32.3
streamlit==1.37.1
streamlit_chat==0.1.1
transformers==4.45.1
tiktoken
torch==1.12.0
accelerate==0.23.0
Werkzeug==3.0.3
watchdog==4.0.2
